requirements.txt
ffmpeg
numpy
pillow
argparse